Attempt to improve the thread pool and clean up the code.
Extracted the thread safe queue in a separate class, worker that represents thread is also
extracted as a class. Added waiting functionality. Also work stealing because some workers might
be idle while others are busy, which by the way in general should not be the case because every
task is assigned to workers using uniform distribution.